Move node draining test from actuator to infra
Node draining is now part of the machine controller.
